Move UI class to a separate file
