+++ VisualStudio/fheroes2/sources.props
@@ -178,6 +178,7 @@
     <ClCompile Include="src\fheroes2\gui\ui_map_object.cpp" />
     <ClCompile Include="src\fheroes2\gui\ui_monster.cpp" />
     <ClCompile Include="src\fheroes2\gui\ui_scrollbar.cpp" />
+    <ClCompile Include="src\fheroes2\gui\ui_slider.cpp" />
     <ClCompile Include="src\fheroes2\gui\ui_option_item.cpp" />
     <ClCompile Include="src\fheroes2\gui\ui_text.cpp" />
     <ClCompile Include="src\fheroes2\gui\ui_tool.cpp" />
@@ -396,6 +397,7 @@
     <ClInclude Include="src\fheroes2\gui\ui_monster.h" />
     <ClInclude Include="src\fheroes2\gui\ui_object_rendering.h" />
     <ClInclude Include="src\fheroes2\gui\ui_scrollbar.h" />
+    <ClInclude Include="src\fheroes2\gui\ui_slider.h" />
     <ClInclude Include="src\fheroes2\gui\ui_option_item.h" />
     <ClInclude Include="src\fheroes2\gui\ui_text.h" />
     <ClInclude Include="src\fheroes2\gui\ui_tool.h" />
